+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- Fixed `derive` not getting called when an origin was set by its own subscription in response to `reflect` ([#18](https://github.com/PixievoltNo1/svelte-writable-derived/issues/18))
 - TypeScript: Any stores you pass to `writableDerived` that `reflect` will be setting must now include a `set` method (and stores passed to `propertyStore` no longer need an `update` method). ([#19](https://github.com/PixievoltNo1/svelte-writable-derived/issues/19))
 
 ## 2.1.3 (March 20, 2022)